# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kryv live streaming platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: Browse on-demand videos (Kryv Watch), optionally filtered by channel, genre, or search
      parameters:
        - name: channelId
          in: query
          required: false
          schema:
            type: integer
        - name: categorySlug
          in: query
          required: false
          schema:
            type: string
        - name: contentType
          in: query
          required: false
          schema:
            type: string
            enum: [upload, original]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoSummary"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Create a video record and start a direct upload session for the current user's channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Created video with an upload URL for the raw video file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoCreateResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Current user has no channel yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a single video's detail and register a view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoDetail"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateVideo
      tags: [videos]
      summary: Update a video's metadata (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoUpdate"
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the video owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the video owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    VideoSummary:
      type: object
      required:
        [
          id,
          title,
          thumbnailUrl,
          posterUrl,
          backdropUrl,
          durationSeconds,
          viewCount,
          channelId,
          channelName,
          channelAvatarUrl,
          categoryId,
          categoryName,
          contentType,
          playbackId,
          uploadStatus,
          createdAt,
        ]
      properties:
        id: { type: integer }
        title: { type: string }
        thumbnailUrl:
          description: Landscape 16:9 thumbnail — used in Kryv Watch grids
          type: ["string", "null"]
        posterUrl:
          description: Portrait 2:3 poster — used in Kryv Cinema rows
          type: ["string", "null"]
        backdropUrl:
          description: Wide cinematic backdrop — used in Kryv Cinema hero banners
          type: ["string", "null"]
        durationSeconds: { type: ["integer", "null"] }
        viewCount: { type: integer }
        channelId: { type: integer }
        channelName: { type: string }
        channelAvatarUrl: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        contentType:
          type: string
          enum: [upload, original]
        playbackId:
          description: Mux playback id for HLS playback. Null until Mux finishes processing the upload.
          type: ["string", "null"]
        uploadStatus:
          type: string
          enum: [waiting, processing, ready, errored]
        createdAt: { type: string, format: date-time }
    VideoDetail:
      allOf:
        - $ref: "#/components/schemas/VideoSummary"
        - type: object
          required: [description, isOwner]
          properties:
            description: { type: ["string", "null"] }
            isOwner: { type: boolean }
    VideoInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 5000 }
        categoryId: { type: integer }
        contentType:
          type: string
          enum: [upload, original]
    VideoCreateResponse:
      allOf:
        - $ref: "#/components/schemas/VideoDetail"
        - type: object
          required: [uploadUrl]
          properties:
            uploadUrl:
              description: Mux direct-upload URL — PUT the raw video file here from the browser
              type: string
    VideoUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 5000 }
        categoryId: { type: integer }
...