# lib/api-zod/src/generated/types/videoCreateResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Kryv live streaming platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoDetail } from './videoDetail';

export type VideoCreateResponse = VideoDetail & {
  /** Mux direct-upload URL — PUT the raw video file here from the browser */
  uploadUrl: string;
};