# lib/api-zod/src/generated/types/videoSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Kryv live streaming platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoSummaryContentType } from './videoSummaryContentType';
import type { VideoSummaryUploadStatus } from './videoSummaryUploadStatus';

export interface VideoSummary {
  id: number;
  title: string;
  /**
     * Landscape 16:9 thumbnail — used in Kryv Watch grids
     * @nullable
     */
  thumbnailUrl: string | null;
  /**
     * Portrait 2:3 poster — used in Kryv Cinema rows
     * @nullable
     */
  posterUrl: string | null;
  /**
     * Wide cinematic backdrop — used in Kryv Cinema hero banners
     * @nullable
     */
  backdropUrl: string | null;
  /** @nullable */
  durationSeconds: number | null;
  viewCount: number;
  channelId: number;
  channelName: string;
  /** @nullable */
  channelAvatarUrl: string | null;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName: string | null;
  contentType: VideoSummaryContentType;
  /**
     * Mux playback id for HLS playback. Null until Mux finishes processing the upload.
     * @nullable
     */
  playbackId: string | null;
  uploadStatus: VideoSummaryUploadStatus;
  createdAt: Date;
}